[UPDATE][PR REVIEW] SINGN-UP&LOGIN VIEW 1. VALIDATION을 위한 함수들을 별도로 모듈화함 2. 불필요한 코드 리팩토링(예. User.objects.get() 3. 잘못 작성된 login  decorator 수정 4. `if utils.password_check() 문을 사용하지 않고 함수 호출 바로 시행 5. 어떠한 토큰을 사용하였는지 구체화하기 위해 변수명 재설정(login_token)
