+++ user/views.py
@@ -7,7 +7,8 @@ from django.http    import JsonResponse
 from django.views   import View
 
 from user.models    import User
-from decorators     import utils
+from user           import utils
+
 from my_settings    import SECRET_KEY, ALGORITHM
 
 
@@ -24,8 +25,7 @@ class SignUpView(View):
             if not utils.account_validation(account): 
                 return JsonResponse({"MESSAGE":'NOT VALID ACCOUNT'}, status=400)
 
-            if utils.password_check(password)!=True:
-                return utils.password_check(password)
+            utils.password_check(password)
 
             if not utils.email_validation(email):
                 return JsonResponse({"MESSAGE":'NOT VALID EMAIL'}, status=400)
@@ -59,22 +59,24 @@ class SignUpView(View):
 class LoginView(View):
     def post(self, request):
         try:
-            data        = json.loads(request.body)
-            signup_db   = User.objects.all()
-            password    = data['password']
-            
-            if not signup_db.filter(account=data['account']).exists():
-                return JsonResponse({'MESSAGE':'ACCOUNT NOT FOUND'}, status=404)
+            data      = json.loads(request.body)
+            account   = data['account']
+            password  = data['password']
 
-            if signup_db.filter(account = data['account']).exists():
-                login_id = User.objects.get(account = data['account'])
+            user           = User.objects.get(account=account)
+            password_in_db = user.password
 
-                if bcrypt.checkpw(password.encode('utf-8'), login_id.password.encode('utf-8')):
-                    user_token = jwt.encode({'user_id': login_id.id}, SECRET_KEY, algorithm=ALGORITHM)
-                    return JsonResponse({'TOKEN':user_token}, status=200)
-                return JsonResponse({'MESSAGE':'INVALID ACCOUNT'}, status= 401)
+
+            if bcrypt.checkpw(password.encode('utf-8'), password_in_db.encode('utf-8')):
+                payload = {"user_id":user.id}
+                login_token = jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
+                return JsonResponse({'MESSAGE':'SUCCESS','TOKEN':login_token}, status=200, safe=True)
+            return JsonResponse({'MESSAGE':'INVALID ACCOUNT'}, status= 401)
 
         except KeyError :
             return JsonResponse({'MESSAGE':'KEY ERROR'}, status=400)
         except ValueError:
-            return JsonResponse({'MESSAGE':'VALUE ERROR'}, status=400)
+            return JsonResponse({'MESSAGE':'VALUE ERROR'},status=400)
+        except User.DoesNotExist:
+            return JsonResponse({"error": "ACCOUNT NOT EXISTS."}, status=404)
+
